KALYAN/requirements.txt
flask
tensorflow
numpy
pillow
